move lutron_db to aiolip
