+++ homeassistant/components/lutron/entity.py
@@ -8,8 +8,8 @@ from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity import Entity
 
 from . import LutronController
+from .aiolip import Device, KeypadComponent, Output, Sysvar
 from .const import DOMAIN
-from .lutron_db import Device, KeypadComponent, Output, Sysvar
 
 
 class LutronBaseEntity(Entity):
