+++ homeassistant/components/lutron/event.py
@@ -10,8 +10,8 @@ from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 from homeassistant.util import slugify
 
 from . import ATTR_ACTION, ATTR_FULL_ID, ATTR_UUID, DOMAIN, LutronController, LutronData
+from .aiolip import Button
 from .entity import LutronKeypadComponent
-from .lutron_db import Button
 
 
 class LutronEventType(StrEnum):
