+++ homeassistant/components/lutron/light.py
@@ -19,9 +19,9 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import CONF_USE_RADIORA_MODE, DOMAIN, LIPLedState, LutronController, LutronData
+from .aiolip import Led, Output
 from .const import CONF_DEFAULT_DIMMER_LEVEL, DEFAULT_DIMMER_LEVEL
 from .entity import LutronKeypadComponent, LutronOutput
-from .lutron_db import Led, Output
 
 _LOGGER = logging.getLogger(__name__)
