+++ homeassistant/components/lutron/sensor.py
@@ -10,8 +10,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
 from . import DOMAIN, LutronController, LutronData
+from .aiolip import Sysvar
 from .entity import LutronVariable
-from .lutron_db import Sysvar
 
 _LOGGER = logging.getLogger(__name__)
